requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

sqlalchemy==2.0.25
python-dotenv==1.0.1

openai==1.14.2
requests==2.31.0

streamlit==1.32.0

passlib==1.7.4
bcrypt==4.1.2
